Unit test for interactor
